ci: Retry failed integration tests (#12260)
Integration tests are inherently flakey. Our test suite has about 15% flake rate
over the past month. It's not a good use of time right now to shave a few
percentages off of that.

Instead, we can just retry the failed tests. To do this, we need to introduce a
new 'rerun' workflow that can be triggered by the integraion test suite on
failure. This rerun workflow then waits for the integration workflow to complete
before triggering a rerun of failed jobs.