update governance to introduce a director role
As requested by the TOC, this PR updates Linkerd's governance to introduce a
director role, which is initially William Morgan. This is a non-technical
leadership role with voting power for decisions.

It also clarifies the decision-making process for steering committee members:
once nominated, there is also a vote to determine inclusion. (Previously
nominations themselves were technically sufficient until the participation limit
was hit.)

Signed-off-by: William Morgan <william@buoyant.io>