Don't patch httproute status if it hasn't changed (#12215)
Fixes #12104

The leader of the policy controllers (as determined by a lease resource) is in charge of keeping the HttpRoute resource statuses up to date.  To accomplish this, it performs reconciliation every 10 seconds where it computes the desired status for each HttpRoute and patches the resource.  When there are many HttpRoute resources in the cluster, most of which change infrequently, this results in a large number of redundant no-op patches: one per HttpRoute every 10 seconds.

These patches are stored in an in-memory unbounded queue while waiting to be dispatched to the Kube API.  If the rate that these patches are generated exceeds the rate that the Kube API can process them (synchronously) then this queue will grow without bound, resulting in a memory leak.

To address this we do two things:
* Use a bounded queue of size 10000. If the queue fills up, patches will be dropped.  Statuses will be reconciled in a later iteration of the reconciliation loop once there is queue capacity.
* Only generate a patch when the desired status differs from the current status.  This greatly reduces the number of patches generated.

Signed-off-by: Alex Leong <alex@buoyant.io>