feat(cli): Add a 'diagnostics profile' command (#12383)
We don't have easy way to inspect the profile information for a given address.

This commit adds a new 'diagnostics profile' command to the CLI that always
returns JSON-formatted profile information for the given address using the
internal protobuf JSON representation. While this isn't necessarily the most
user-friendly interface to this data, it gives us an unfiltered view of the API
response, which is typically what we want when debugging.