Policy controller should index services by clusterIPs
This changes the policy controller service indexer to index
services by thier `spec.clusterIPs` instead of `spec.clusterIP`, to
account for dual-stack services that hold both an IPv4 and an IPv6
address in `spec.clusterIPs`.

This change is fully backwards-compatible.

This assists the ongoing effort to support IPv6/dual-stack in linkerd,
but doesn't imply full support yet.