Fix destination staleness issue when adding EndpointSlices
When updating the portPublisher's address set when a new EndpointSlice
creation event is received, its addresses where getting overwritten with
stale data whenever its IDs already existed in the current pp's address
set.

This bug would be manifested for example when dealing with dual-stack
services. We'd have two ES with different IPs but with the same
TargetRef pointing to the same pod, so their IDs would be the same. When
the creation event for the second ES is received, its addresses would
get overridden with the ones from the first ES.

Note however that dual-stack services are not supported yet, but we have
unit tests for them in `server_test.go` although this particular failure
case wasn't covered. Upcoming dual-stack support will modify this logic
a bit so there's no need to add that case.

Nevertheless, there can also be pathological cases in single-stack where
this can be a problem. For example when ES get recycled but the deletion
event is not caught for some reason, when the addition event is received
its Address data will be overwritten by the old stale entry.

## Other Changes

- Remove overriding `newAddressSet.LocalTrafficPolicy` as that is
  already taken care inside `pp.endpointSliceToAddresses(slice)`.
- When there are no Add events to send, return early without updating
  state nor updating metrics.
- Finally, the IPv6 address for the "name1-ipv6" ES test fixture was
updated to match the associated pod.