Add IPv6 support for the destination controller
Services in dual-stack mode result in the creation of two
EndpointSlices, one for each IP family. Before this change, the Get
Destination API would nondeterministically return the address for any of
those ES, depending on which one was processed last by the controller
because they would overwrite each other.

As part of the ongoing effort to support IPv6/dual-stack networks, this
change fixes that behavior giving preference to IPv6 addresses whenever
a service exposes both families.

Also the server unit tests were updated to segregate the tests and
resources dealing with the IPv4/IPv6/dual-stack cases.