feat(inject): support arbitrary proxy parameters
New proxy HTTP2 server parameters can be configured from the environment. These
environment variables take the form:

    LINKERD2_PROXY_[INBOUND|OUTBOUND]_[SERVER]_[HTTP2]_<PARAM>

We want to allow these parameters to be set, i.e., in a values.yaml file without
having to document all of the possible parameters (especially those that
primarily influence internals not exposed to the application).

To accomodate this, this commit introduces generic proxy environment templating
and a test to verify the behavior.

Values like the following:

    proxy:
      inbound:
        scope:
          proto:
            blueberryPi: 3.14
      outbound:
        scope:
          proto:
            applesauce: "valueA"

Manifest as a proxy environment including:

    LINKERD2_PROXY_INBOUND_SCOPE_PROTO_BLUEBERRY_PI="3.14"
    LINKERD2_PROXY_OUTBOUND_SCOPE_PROTO_APPLESAUCE=valueA

That is, Helm value keys are converted to SCREAMING_SNAKE_CASE environment
variable names.