Update to latest version of kube_codegen.sh
We use the `kube_codegen.sh` script from the kubernetes/code-generator
repo to generate Go bindings for CRDs defined in this repo. The
arguments and flags for that script changed in Kubernetes 1.30, which
broke our usage of that script when this repo's Kubernetes dependencies
were updated in #12463.

In this change, I'm updating `bin/update-codegen.sh` to work with the
latest version of the `kube_codegen.sh` script.

To validate, check out this branch and run `bin/update-codegen.sh`. The
script should exit cleanly, and there should be no changes to the
previously-generated Go bindings.

Signed-off-by: Kevin Ingelman <ki@buoyant.io>