chore!: Update the minimum Kubernetes version to 1.25
Kubernetes 1.25 introduces and 1.29 stabilizes a new CRD validation feature.
This feature is used by newer versions of the Gateway API CRDs. The Gateway API
maintains compatibility with "at least the last 5" Kubernetes minor releases.
Now that 1.30 has been released, they have dropped support for versions 1.24 and
older.

We are going to have to update our Minimum Kubernetes version in the near future
anyway, as we update our kube-rs dependency.

This change updates the minimum Kubernetes version to 1.25 in advance of these
upcoming changes.

[rules]: https://kubernetes.io/blog/2022/09/23/crd-validation-rules-beta/