Set backend_not_found route status when any backends are not found (#12565)
The policy controller sets a status on HttpRoute resources depending on the state of the backends.  If all the backends can be resolved, it should set a resolved_ref status and if any of the backends cannot be resolved, it should set a backend_not_found status.

We update the logic to reflect this.  Additionally, this resolves the edge case where an empty list of backends was resulting in a backend_not_found status instead of a resolved_refs status.

Signed-off-by: Alex Leong <alex@buoyant.io>