Revert "Fix destination staleness issue when adding EndpointSlices (#12427)" (#12589)
This reverts commit 4fccf3e9eccb8179b802725e816e423a3d3b824d.

The early return was causing `pp.addresses = newAddressSet` to not be run when the list of addresses is empty; but setting that is still necessary so that labels are tracked correctly.

This was caught by the tap (viz) integration test run in the release workflow.

(cherry picked from commit 9bd8c005dab06e373b645d5091789d4e4441e593)