Add support for json output in inject and uninject commands (#12600)
We add support for the `--output/-o` flag in `linkerd inject` and `linkerd uninject` commands.  The supported output formats are yaml (default) and json.  Kubectl is able to accept both of these formats which means that `linkerd inject` and `linkerd uninject` output can be piped into kubectl regardless of which output format is used.

Signed-off-by: Alex Leong <alex@buoyant.io>