Allow control plane components to specify concurrency
Concurrency in Linkerd is configured globally through a `proxy.cores`
value or it can be configured through the injector. Control plane
components cannot be configured through the injector since they come
with the proxy pre-configured. In certain cases, it is necessary for
proxies to specify their concurrency. Even though a global value exists,
that directly influences all workloads in the cluster which can have
other implications.

Each control plane component has a `<component>ProxyResources` field
that is used to configure limits and requests. We overload this field
and allow `cores` to be specified. We do not expect this value to be
used often since directly specifying the number of cores might not make
a lot of sense -- this does offer an escape hatch when the control plane
components can't scale up to meet bursty workloads.

Signed-off-by: Matei David <matei@buoyant.io>