Fix issue where initial outbound policy did not contain producer routes (#12619)
An HTTPRoute whose parentRef is a Service in the same namespace is called a producer route. Producer routes should be used in outbound policy by all clients calling that Service, even if the client is in a different namespace.  The policy controller has a bug where when a outbound policy watch is started, the initial outbound policy returned will not include any producer routes which already exist.

We correct this bug and add tests.

Signed-off-by: Alex Leong <alex@buoyant.io>