Make logHTTPHeaders optional (#12747)
The `proxy.logHTTPHeaders` Helm variable supports two values: "insecure" and "off".  However, if this value is absent entirely, validation will fail because it it not one of the supported values.  This can cause issues with backwards compatibility.

We make this value optional by adding "" as a legal value during validation.  Leaving this value unset is equivalent to "off" which is the default value.

Signed-off-by: Alex Leong <alex@buoyant.io>