add ability to read prometheus basic auth creds from secret (#13035)
If a Prometheus instance is configured with basic auth, the linkerd-viz metric-api is unable to access it.  It is possible to provide basic auth credentials in the `prometheusUrl` Helm value, but this leaks the credentials into the deployment manifest.

We add a Helm value called `prometheusCredsSecret` to the Linkerd-viz helm chart.  When set, this causes the metrics-api to read `user` and `password` fields from a secret with the given name in the Linkerd-viz namespace and to use these values for basic auth when making requests to Prometheus.

These credentials are loaded at startup and any changes to the credentials require a restart of the metrics-api to take effect.

Signed-off-by: Alex Leong <alex@buoyant.io>