Fix pod annotations propagating to namespace
Currently, there's a bug with the jaeger injector where annotation overrides from a pod can propagate to any pod in the same namespace, provided they happen after the admission for the original pod occurs.

This fixes that by performing a deep copy of the namespace annotations, instead of editing the one in the local cache (which appears to have been unintentional).

Validated with a local deployment and checking that annotations are no longer propagated to unrelated pods.

Signed-off-by: Scott Fleener <scott@buoyant.io>