Allow diagnostics endpoints command to receive more than one message (#13285)
The `linkerd diagnostics endpoints` command initiates a `Get` lookup to the destination controller to get the set of endpoints for a destination.  This is a streaming response API and the command takes only the first response message and displays it.  However, the full current state of endpoints may be split across multiple messages, resulting in an incomplete list of endpoints displayed.

We instead read continuously from the response stream for a short amount of time (5 seconds) before displaying the full set of endpoints received.

Signed-off-by: Alex Leong <alex@buoyant.io>